Fix "locus not found" error when using foreach
Add locus id as argument to filter.params() function and add this
argument to all calls to filter.params() in analysis_functions.R.

When filter.params() is called within a foreach loop, locus does not
appear to be passed to the function (despite being an environment).
This results in the error: object 'locus' not found. Fix this by
adding locus id as an argument to filter.params().