updated run.univ
changed the conditional outputting of the h2.latent column in run.univ to depend on whether or not there is a 'prevalence' column in the input.info file rather than whether there are any binary phenotypes in the locus object and whether all values are non-na's. Reason being that for some loci the estimated h2.latent might be NA even if there is a prevalence estimate provided, leading to differences in the number of columns in the univ output for different runs. The reason the if(any(locus$binary)) check was also removed was because its possible that if a mix of binary and continuous phenotypes are analysed, that some binary phenos might not be present in all loci, which could also lead to inconsistencies in the number of columns outputted by this function across loci.

So now it will just always output the h2.latent for all loci whenever there is a prevalence column, regardless if the analysed locus contains binary phenotypes or non-na values in the locus$h2.latent vector